Add RxSwift, SnapKit and RIBs frameworks via Podfile.
Summary:
Use Cocoapods to add dependencies:
- RxSwift
- SnapKit
- RIBs
for the main app target, as well as
- RxTest
- Sourcery
- Quick
- Nimble
for the test target.

Also, add .gitignore to not add Pods directory to the source control.
For the purposes of this tutorial that's fine, in the actual production environment, most likely, committing all dependencies into git would be a better approach.