Add RIBs Xcode templates.
Summary:
RIBs framework comes with a set of Xcode templates that simplify adding RIBs to the project by generating necessary protocol/class declarations that follow a specific pattern.
It is a good practice to add these templates to the project repo (to make the project senf-contained), moreover, the Xcode template is not part of the corresponding Cocoapod.