Add 4 first unit-tests for Routers and Interactors.
Summary:
Now that we have protocol mocks, we can instantiate RootRouter, RootInteractor, SplashRouter, SplashInteractor by providing them with "fake" (mock) dependencies, instead of actual objects.
I.e., to instantuate `RootRouter` object in a unit-test, we provide it with mock implementations of `interactor: RootInteractable`, `viewController: RootViewControllable` and `splashBuilder: SplashBuildable` objects,
so that we can test side-effects easily.

There is one unsolved problem still: base RIB's protocols aren't annotated with `/// sourcery: ` annotations, and the mock generator did not generate them for us.
These are required to be able to instantiate instances of *mock* classes, e.g., `RootRoutingMock` - it requires `interactable: Interactable` and `viewControllable: ViewControllable` objects to be passed in its initializer.
We'll handle this in the next commit.